+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/chapters/ChaptersFragment.kt
@@ -15,6 +15,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.compose.ui.platform.rememberNestedScrollInteropConnection
 import androidx.fragment.app.activityViewModels
@@ -24,12 +26,21 @@ import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.models.to.Chapter
 import au.com.shiftyjelly.pocketcasts.player.view.PlayerContainerFragment
+import au.com.shiftyjelly.pocketcasts.player.view.chapters.ChaptersViewModel.NavigationState
+import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingFlow
+import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingLauncher
+import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingUpgradeSource
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
+import au.com.shiftyjelly.pocketcasts.ui.theme.ThemeColor
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureTier
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
 import au.com.shiftyjelly.pocketcasts.views.helper.UiUtil
+import com.google.android.material.snackbar.Snackbar
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.collectLatest
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @AndroidEntryPoint
@@ -52,6 +63,8 @@ class ChaptersFragment : BaseFragment() {
                 val uiState by chaptersViewModel.uiState.collectAsStateWithLifecycle()
                 val lazyListState = rememberLazyListState()
                 this@ChaptersFragment.lazyListState = lazyListState
+                val context = LocalContext.current
+                val currentView = LocalView.current
 
                 val scrollToChapter by chaptersViewModel.scrollToChapterState.collectAsState()
                 LaunchedEffect(scrollToChapter) {
@@ -65,13 +78,38 @@ class ChaptersFragment : BaseFragment() {
                     }
                 }
 
+                LaunchedEffect(Unit) {
+                    chaptersViewModel
+                        .navigationState
+                        .collectLatest { event ->
+                            when (event) {
+                                is NavigationState.StartUpsell -> startUpsell()
+                            }
+                        }
+                }
+
+                LaunchedEffect(Unit) {
+                    chaptersViewModel
+                        .snackbarMessage
+                        .collectLatest { message ->
+                            Snackbar.make(currentView, context.getString(message), Snackbar.LENGTH_SHORT)
+                                .setBackgroundTint(ThemeColor.playerContrast01(Theme.ThemeType.DARK))
+                                .setTextColor(ThemeColor.playerBackground01(Theme.ThemeType.DARK, theme.playerBackgroundColor(uiState.podcast)))
+                                .show()
+                        }
+                }
+
                 Surface(modifier = Modifier.nestedScroll(rememberNestedScrollInteropConnection())) {
                     ChaptersPage(
                         lazyListState = lazyListState,
-                        chapters = uiState.chapters,
+                        chapters = uiState.displayChapters,
+                        totalChaptersCount = uiState.totalChaptersCount,
                         onSelectionChange = { selected, chapter -> chaptersViewModel.onSelectionChange(selected, chapter) },
                         onChapterClick = ::onChapterClick,
                         onUrlClick = ::onUrlClick,
+                        onSkipChaptersClick = { chaptersViewModel.onSkipChaptersClick(it) },
+                        isTogglingChapters = uiState.isTogglingChapters,
+                        showSubscriptionIcon = uiState.showSubscriptionIcon,
                         backgroundColor = uiState.backgroundColor,
                     )
                 }
@@ -100,5 +138,15 @@ class ChaptersFragment : BaseFragment() {
 
     private fun showPlayer() {
         (parentFragment as? PlayerContainerFragment)?.openPlayer()
+    }
+
+    private fun startUpsell() {
+        val source = OnboardingUpgradeSource.SKIP_CHAPTERS
+        val onboardingFlow = OnboardingFlow.Upsell(
+            source = source,
+            showPatronOnly = Feature.DESELECT_CHAPTERS.tier == FeatureTier.Patron ||
+                Feature.DESELECT_CHAPTERS.isCurrentlyExclusiveToPatron(),
+        )
+        OnboardingLauncher.openOnboardingFlow(activity, onboardingFlow)
     }
 }