+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/entity/BaseEpisode.kt
@@ -41,6 +41,7 @@ sealed interface BaseEpisode {
     var playedUpToModified: Long?
     var playErrorDetails: String?
     var downloadErrorDetails: String?
+    var deselectedChapters: ChapterIndices
 
     // temporary variables
     var playing: Boolean