+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/EpisodeManagerImpl.kt
@@ -14,6 +14,7 @@ import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedNumbers
 import au.com.shiftyjelly.pocketcasts.models.db.helper.LongestEpisode
 import au.com.shiftyjelly.pocketcasts.models.db.helper.YearOverYearListeningTime
 import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
+import au.com.shiftyjelly.pocketcasts.models.entity.ChapterIndices
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
@@ -351,6 +352,24 @@ class EpisodeManagerImpl @Inject constructor(
             episodeDao.updateAutoDownloadStatus(autoDownloadStatus, episode.uuid)
         } else {
             userEpisodeDao.updateAutoDownloadStatus(autoDownloadStatus, episode.uuid)
+        }
+    }
+
+    override suspend fun updateAutomaticallyCachedStatus(
+        episode: BaseEpisode?,
+        automaticallyCached: Boolean,
+    ) {
+        if (episode == null || episode is UserEpisode) return
+        (episode as PodcastEpisode).isAutomaticallyCached = automaticallyCached
+        episodeDao.updateAutomaticallyCachedStatus(automaticallyCached, episode.uuid)
+    }
+
+    override suspend fun cleanAutomaticallyCachedEpisodes(playbackManager: PlaybackManager) {
+        val episodes = episodeDao.findAutomaticallyCachedEpisodes()
+        episodes?.forEach { episode ->
+            episode.isAutomaticallyCached = false
+            updateAutomaticallyCachedStatus(episode, automaticallyCached = false)
+            deleteEpisodeFile(episode, playbackManager, disableAutoDownload = true, removeFromUpNext = false)
         }
     }
 
@@ -1098,4 +1117,18 @@ class EpisodeManagerImpl @Inject constructor(
 
             return@withContext newDownloadUrl ?: episode.downloadUrl
         }
+
+    override suspend fun selectChapterIndexForEpisode(chapterIndex: Int, episode: PodcastEpisode) {
+        val deselectedChapterIndices = episode.deselectedChapters
+        if (!deselectedChapterIndices.contains(chapterIndex)) return
+        episode.deselectedChapters = ChapterIndices(deselectedChapterIndices - chapterIndex)
+        episodeDao.update(episode)
+    }
+
+    override suspend fun deselectChapterIndexForEpisode(chapterIndex: Int, episode: PodcastEpisode) {
+        val deselectedChapterIndices = episode.deselectedChapters
+        if (deselectedChapterIndices.contains(chapterIndex)) return
+        episode.deselectedChapters = ChapterIndices(deselectedChapterIndices + chapterIndex)
+        episodeDao.update(episode)
+    }
 }