+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/sync/update/SyncUpdateResponse.kt
@@ -13,6 +13,7 @@ import com.pocketcasts.service.api.SyncUserEpisode
 import com.pocketcasts.service.api.SyncUserFolder
 import com.pocketcasts.service.api.SyncUserPlaylist
 import com.pocketcasts.service.api.SyncUserPodcast
+import com.pocketcasts.service.api.UserPodcastResponse
 import com.pocketcasts.service.api.addToUpNextOrNull
 import com.pocketcasts.service.api.addToUpNextPositionOrNull
 import com.pocketcasts.service.api.autoArchiveEpisodeLimitOrNull
@@ -134,6 +135,45 @@ data class SyncUpdateResponse(
                     showArchived = syncUserPodcast.settings.showArchived?.value?.value,
                     showArchivedModified = syncUserPodcast.settings.showArchived?.modifiedAt?.toDate(),
                 )
+
+            fun fromUserPodcastResponse(response: UserPodcastResponse) = PodcastSync(
+                uuid = response.uuid,
+                subscribed = true,
+                episodesSortOrder = response.settings.episodesSortOrderOrNull?.value?.value,
+                folderUuid = response.folderUuidOrNull?.value,
+                sortPosition = response.sortPositionOrNull?.value,
+                dateAdded = response.dateAddedOrNull?.toDate(),
+                startFromSecs = response.settings.autoStartFromOrNull?.value?.value,
+                startFromModified = response.settings.autoStartFromOrNull?.modifiedAt?.toDate(),
+                skipLastSecs = response.settings.autoSkipLastOrNull?.value?.value,
+                skipLastModified = response.settings.autoSkipLastOrNull?.modifiedAt?.toDate(),
+                addToUpNext = response.settings.addToUpNextOrNull?.value?.value,
+                addToUpNextModified = response.settings.addToUpNextOrNull?.modifiedAt?.toDate(),
+                addToUpNextPosition = response.settings.addToUpNextPositionOrNull?.value?.value,
+                addToUpNextPositionModified = response.settings.addToUpNextPositionOrNull?.modifiedAt?.toDate(),
+                useCustomPlaybackEffects = response.settings.playbackEffectsOrNull?.value?.value,
+                useCustomPlaybackEffectsModified = response.settings.playbackEffectsOrNull?.modifiedAt?.toDate(),
+                playbackSpeed = response.settings.playbackSpeedOrNull?.value?.value,
+                playbackSpeedModified = response.settings.playbackSpeedOrNull?.modifiedAt?.toDate(),
+                trimSilence = response.settings.trimSilenceOrNull?.value?.value,
+                trimSilenceModified = response.settings.trimSilenceOrNull?.modifiedAt?.toDate(),
+                useVolumeBoost = response.settings.volumeBoostOrNull?.value?.value,
+                useVolumeBoostModified = response.settings.volumeBoostOrNull?.modifiedAt?.toDate(),
+                showNotifications = response.settings.notificationOrNull?.value?.value,
+                showNotificationsModified = response.settings.notificationOrNull?.modifiedAt?.toDate(),
+                autoArchive = response.settings.autoArchiveOrNull?.value?.value,
+                autoArchiveModified = response.settings.autoArchiveOrNull?.modifiedAt?.toDate(),
+                autoArchivePlayed = response.settings.autoArchivePlayedOrNull?.value?.value,
+                autoArchivePlayedModified = response.settings.autoArchivePlayedOrNull?.modifiedAt?.toDate(),
+                autoArchiveInactive = response.settings.autoArchiveInactiveOrNull?.value?.value,
+                autoArchiveInactiveModified = response.settings.autoArchiveInactiveOrNull?.modifiedAt?.toDate(),
+                autoArchiveEpisodeLimit = response.settings.autoArchiveEpisodeLimitOrNull?.value?.value,
+                autoArchiveEpisodeLimitModified = response.settings.autoArchiveEpisodeLimitOrNull?.modifiedAt?.toDate(),
+                episodeGrouping = response.settings.episodeGroupingOrNull?.value?.value,
+                episodeGroupingModified = response.settings.episodesSortOrderOrNull?.modifiedAt?.toDate(),
+                showArchived = response.settings.showArchived?.value?.value,
+                showArchivedModified = response.settings.showArchived?.modifiedAt?.toDate(),
+            )
         }
     }